Fixing crud for posts

+++ API/Controllers/PostsController.cs
@@ -1,7 +1,9 @@
 using API.Entities;
+using API.Extensions;
 using API.Interfaces;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace API.Controllers
 {
@@ -15,9 +17,11 @@ namespace API.Controllers
             _postRepository = postRepository;
 
         }
-        [HttpGet("{username}")]
-        public async Task<ActionResult<IEnumerable<Post>>> GetPostsForUser(string username)
+        [HttpGet]
+        public async Task<ActionResult<IEnumerable<Post>>> GetPostsForUser()
         {
+            var username = User.GetUsernameFromToken();
+            if(username == null) return NotFound("user not found");
             var user = await _userManager.FindByNameAsync(username);
             if(user == null) return NotFound("user not found");
             var posts = await _postRepository.GetPostsForUser(username);
@@ -25,7 +29,14 @@ namespace API.Controllers
         }
         [HttpPost]
         public async Task<ActionResult> AddPost(Post post){
-           await _postRepository.AddPost(post);
+             var username = User.GetUsernameFromToken();
+            var user = await _userManager.FindByNameAsync(username);
+            if(user == null) return NotFound("User not found");
+           await _postRepository.AddPost(new Post{
+            AuthorId = user.Id,
+            TextContent = post.TextContent,
+            ImgUrl = post.ImgUrl
+           });
            if(await _postRepository.SaveChangesAsync()){
             return Ok("Succesfully added new post");
            }else{
@@ -35,11 +46,29 @@ namespace API.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeletePost(int id)
         {
+               var username = User.GetUsernameFromToken();
+            var user = await _userManager.Users.Include(p => p.Posts).FirstOrDefaultAsync(u => u.UserName == username);
+            var userpost = user.Posts.FirstOrDefault(p => p.Id == id);
+            if(userpost == null) return BadRequest("Unable to delete");
            await _postRepository.DeletePost(id);
            if(await _postRepository.SaveChangesAsync()){
             return Ok("Succesfully deleted post");
            }else{
             return BadRequest("Failed to delete post");
+           }
+        }
+        [HttpPut("{id}")]
+        public async Task<ActionResult> EditPost(int id, [FromBody] Post post)
+        {
+             var username = User.GetUsernameFromToken();
+            var user = await _userManager.Users.Include(p => p.Posts).FirstOrDefaultAsync(u => u.UserName == username);
+            var userpost = user.Posts.FirstOrDefault(p => p.Id == id);
+            if(userpost == null) return BadRequest("Unable to edit");
+            await _postRepository.EditPost(id,post);
+           if(await _postRepository.SaveChangesAsync()){
+            return Ok("Succesfully edited post");
+           }else{
+            return BadRequest("Failed to edit post");
            }
         }
     }
